Get wasm build working
This will let us easily deploy the toy on the web.
