+++ splinetoy/Cargo.toml
@@ -4,7 +4,13 @@ version = "0.1.0"
 authors = ["Colin Rofls <colin@cmyr.net>"]
 edition = "2018"
 
+[lib]
+crate-type = ["cdylib", "rlib"]
+
 [dependencies]
 druid = "0.6"
 serde_json = "1.0.60"
 spline = { version = "0.3.0", features=["serde"], path = ".."}
+
+wasm-bindgen = "0.2.67"
+console_error_panic_hook = "0.1.6"
