+++ splinetoy/index.js
@@ -1,0 +1,8 @@
+import init, { wasm_main } from "./pkg/splinetoy.js";
+
+async function run() {
+  await init();
+  wasm_main();
+}
+
+run();
